Add variable names to placeholders.
Placeholders may be written as [VARNAME:PATH.OBJECT]. Anytime another
placeholder references VARNAME, it will refer to the same object for
consitend renderings.

This means that if the full path is the same, the same result is
rendered.

If the path is partially the same, the same objects will be used for the
common path. This enables consistency in the objects used.