Allow passing variables to subobjects.
List of parameters may be defined in objects as:
  <OBJ(PARAM1)>

or
  <OBJ(PARAM1, PARAM2)>

and later used as ordinary named variables in placholders like
  [PARAM1:ROOT.SUBOBJECT]

Since this is a named variable, [PARAM1:ROOT.SUBOBJECT] will now refer
to the same object/string.

Parameters are passed like:
  [ROOT.SUBOBJECT(PARAM1)]

or
  [ROOT.SUBOBJECT(PARAM1, PARAM2)]

If the variable names are already defined, that ObjectInstance will be
passed, otherwise a new is created, which can be used later.

The variable names that are passed in the placeholder may be different
than the parameter names expected in the subobject, but the number must
match.